==> 04-Individuals.Rmd <==
```{r setupindividuals, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(raster)
library(leaflet)
library(sf)
library(lidR)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
```

# Individuals

This chapter support the choice of individuals for the study.

## **Angela** - *Dicorynia guyanensis*
 
```{r angleadrone,  fig.cap="Angela from the drone."}
include_graphics("data/Regina/drone/Angela/DJI_2910.JPG")
```

## *Sextonia rubra*

**Need to report TreeD data here.**

## *Eschweilera sp.*

### MN5505 - Regina

**E. coriacea** following Pascal. Currently the' best candidate near Regina.

```{r MN5505ground, fig.cap="MN5505 from the ground."}
include_graphics("data/Regina/pictures/TreeMutation20210308/MN5505/_0018436.jpg")
```

```{r MN5505drone, fig.cap="MN5505 from the drone."}
include_graphics("data/Regina/drone/MN5505/DJI_1550.JPG")
```
