==> 81-A1Meetings.Rmd <==
# (APPENDIX) Appendix {-} 

# Appendix 1: Meetings

## 11/02/2021 Sylvain's Kick-off

*NT, MH, & SS*

### Funding

* GPR (labex COTE + idex UB): audition 01/21, answer 03-06/21
* Currently enough for __2__ species this year

### Site

* __Regina__ 
    * 1 LiDAR in 2016 (ONF, see Fig. \@ref(fig:reginaMap) and Fig. \@ref(fig:\@ref(fig:reginaMap)))
    * Possibility of clearing for drone
    * Destructive sampling for tree rings? Coring individuals. **Ask ONF.**
    * Higher trees. Maybe not older but at least implying more cell divisions regarding mutations.
* __Paracou__
    * 4 LiDAR campaigns from 2004 to 2015 (Fig. \@ref(fig:paracouMap))
    * Proximity, easier to follow phenology and collect flowers and fruits

### Species

1. *Dicorynia guyanensis*
1. ...
1. *Eschweilera sp*

* _Sextonia rubra_: 
    * already sampled
    * Paracou
* _Dicorynia_: 
    * Paracou & Regina
    * flowering time? **Ask Natalia Norden and Jerome Chave for original data + Patrick Heuret for his shared article**
    * Suggestion: spot one already flowering before sampling
    * Check the phenology at the herbarium
* _Eschwerilera sp_: 
    * Parcou (small) and Regina (big)
* _Handoranthus sp_: 
    * published genosegmenmes (to read) but high variation
    * Regina
    * Need to be collected for flux cytometry
* _Caryocar sp_: 
    * Paracou (big)
    * many lianas
    * no self shading?

### Field work

* __Regina__
    * beginning of march
    * wood samples
    * drone
    * dry leaves of *Handoranthus sp* for flux cytometry from our favorite individuals and two other potential candidates
    * Sample (Big shot) *Eschweilera sp* for identification and gene pool attribution with Myriam's SSRseq experiment (results in spring)
* __Paracou__
    * _Dycorinia_ from interplots. **Ask Partick Heuret and Eric Nicolini.**
    
### Genomics

* Check flux cytometry confidence: plot second against first sampling for confidence envelope (Fig. \@ref(fig:fluxconfidence))
* choose what to sequence
* get in touch asap sequencing company
* High molecular DNA extraction at CNRGV
* Bionano with CNRGV
* PacBio for HiFi reads

## 27/05/2021 William Marrande

### Sampling

* No experience on cambium
* PacBio 50-150kb < Bionano
* CTAB or Kit
* Send by plane

### Assembly

* PacBio HiFi best contiguity
* hifiasm separate haplophase
* different apporaches for optical map
    * calssical assembly with bionano software (only 3% of N recently)
    * if issues (gaps or N), local tools to make it better

## 06/07/2021 Mutation detection

### Virtual experiment

* Barplot of callers performance across AF at low (50X) and high (100X) coverage
* Check `VarScan` AF=0.1 C=200

### Swiss data

* Map swiss mutations (+1000 bp) on 3P genome
* Compare GATK/Strelka2
* Full genome
*  Stringent filters common with both GATK and Strelka2

### Angela

* 

## 21/01/2022 Mid-project meeting

### Notes

* UVs can have a different signal than direct light exposure measured by Accupar or currently modeled with lidar data (up to 20-30% uncertainty according to G. Vincent)
* lidar gave a hieght of 50.6m against 51.25 with tree climbing. This erros below the meter is very encouraging.
* *Herrera CM 2019* on the phenotypic variability within individual trees or plants
* Screwdriver can be good to initiate the auger (O. Martin)

### Follow-up

* **Branch dating**
    * discuss with Patrick and perhaps Romain methodology, sampling points, tools etc
    * plan a mission with climbers
    * send samples to Patrick
* **Direct light measurement**: not primordial to do aside from a mission if time
* **Mutation detection**: continue with the plan
* **Offspring genotyping**: meet with geneticists to discuss samples and techniques, dimensions of the assay.
* **Data management**: discuss archiving with everybody, e.g. NCBI for raw genomic data, Zenodo for cleaned genomic data, GitHub for codes, what for the others (lidar) ?

## 10/02 Carrotage

*Patrick, Niklas, moi*

* Jocelyn, Valentine, Patrick, Niklas?, me?, Olivier Brunaux?
* 21-25/03
* 3 days
* Book climbers
* Book 2 cars
* Book accommodation in St-George or Regina
* OM for me, Niklas and Jocelyn
* Material
    * carottage => Patrick
    * climbing kit, ask Ilo
    * Accupar
    
## 05/05 Sixto's results

*Niklas, Myriam, Thibault, me*

* Globally: don't follow archi, no signal in UVs, and the estimate gives a high # of low AF mutations + evidence the difficulty to estimate mutations # comparing with previous studies
* Next steps: annotation (w transcripts, contact Christophe & Isabelle) & transmission (suggest a number of shared mutations in branches with position in the genome + ranking)
* Meeting: schedule a meeting with everybody to discuss valorisation
* 3 papaers: 2 genomes, 1 mutations
* Save data
* Follow mutations sharing along architecture (cell surfing idea)
* Add architecture when comparing phylogenies
* Add comparisons with oka if possible

## 20-21/03 Mutations days

* Collection Chêne Saint-Malo
* mutasom limitée
* majoritairement neutre
* rare dans la population donc lente à se répandre
* minmap for pangenome graph => Verzy ?
* Karst et al. 2023 overinterpretation myccorhizal
*
